Move Tip component to folder 'Share'
