+++ app/components/Share/Tip.js
@@ -2,19 +2,19 @@ import React, {Component} from 'react';
 import { View, Text, StyleSheet, Image } from 'react-native';
 import DeviceInfo from 'react-native-device-info'
 
-import { LocalizationContext } from './Translations';
-import PressableLabel from './Share/PressableLabel';
+import { LocalizationContext } from '../Translations';
+import PressableLabel from './PressableLabel';
 import { Icon } from 'native-base';
-import Color from '../themes/color';
-import customStyle from '../themes/customStyle';
-import cardListItemStyle from '../themes/cardListItemStyle';
+import Color from '../../themes/color';
+import customStyle from '../../themes/customStyle';
+import cardListItemStyle from '../../themes/cardListItemStyle';
 import { TouchableRipple } from 'react-native-paper';
 
-import { getTipByScreenName } from '../helpers/tip_helper';
-import { cardBorderRadius } from '../constants/border_radius_constant';
-import { getDeviceStyle } from '../utils/responsive_util';
-import TipTabletStyles from '../styles/tablet/TipComponentStyle';
-import TipMobileStyles from '../styles/mobile/TipComponentStyle';
+import { getTipByScreenName } from '../../helpers/tip_helper';
+import { cardBorderRadius } from '../../constants/border_radius_constant';
+import { getDeviceStyle } from '../../utils/responsive_util';
+import TipTabletStyles from '../../styles/tablet/TipComponentStyle';
+import TipMobileStyles from '../../styles/mobile/TipComponentStyle';
 
 const responsiveStyles = getDeviceStyle(TipTabletStyles, TipMobileStyles);
 
@@ -24,7 +24,7 @@ export default class Tip extends Component {
   renderTipIcon() {
     return (
       <View style={[cardListItemStyle.statusIconWrapper, responsiveStyles.tipIconContainer]}>
-        <Image source={require('../assets/images/tip.png')} style={responsiveStyles.tipIcon}/>
+        <Image source={require('../../assets/images/tip.png')} style={responsiveStyles.tipIcon}/>
       </View>
     )
   }
